- Adding getFullName() for DatabaseObject - Workaround for Oracle driver bug described in http://issues.apache.org/jira/browse/DDLUTILS-29?page=all - Workaround for Oracle driver bug where the column precision is returned as a long, not an int
git-svn-id: http://svn.code.sf.net/p/schemacrawler/code@53 3243f2e7-4e1b-42d2-9742-d984016aa316

+++ schemacrawler/src/main/java/schemacrawler/crawl/MutableColumnDataType.java
@@ -72,7 +72,7 @@ final class MutableColumnDataType
 
   private boolean userDefined;
   private int type;
-  private int precision; // maximum precision
+  private long precision; // maximum precision, long to account for the Oracle driver
   private String literalPrefix; // prefix used to quote a literal
   private String literalSuffix; // suffix used to quote a literal
   private String createParameters; // parameters used in creating the type
@@ -242,7 +242,7 @@ final class MutableColumnDataType
   /**
    * @return Returns the precision.
    */
-  public int getPrecision()
+  public long getPrecision()
   {
     return precision;
   }
@@ -366,7 +366,7 @@ final class MutableColumnDataType
    * @param precision
    *          The precision to set.
    */
-  void setPrecision(final int precision)
+  void setPrecision(final long precision)
   {
     this.precision = precision;
   }
